resolved conflict

+++ service-loans/src/main/java/com/capitaworld/service/loans/service/fundseeker/corporate/impl/LoanApplicationServiceImpl.java
@@ -6290,7 +6290,7 @@ public class LoanApplicationServiceImpl implements LoanApplicationService {
 
 		if(!CommonUtils.isObjectNullOrEmpty(productDetails)){
 			List<ApplicationProposalMapping> applicationProposalMappingList = applicationProposalMappingRepository.getListByApplicationIdAndOrgId(proposalDetails.getApplicationId(),loanApplicationRequest.getNpOrgId());
-			if(!CommonUtils.isObjectNullOrEmpty(applicationProposalMappingList)){
+			if(!CommonUtils.isObjectNullOrEmpty(applicationProposalMappingList) && CommonUtils.isObjectNullOrEmpty(proposalDetails.getNbfcFlow())){
 				Integer deleteApplicationData = applicationProposalMappingRepository.deleteByApplicationIdAndOrgId(proposalDetails.getApplicationId(),loanApplicationRequest.getNpOrgId());
 				if(!CommonUtils.isObjectNullOrEmpty(deleteApplicationData) && deleteApplicationData>0){
 					logger.info("Data deleted for applicationId:"+proposalDetails.getApplicationId()+" and for fpProductId:"+proposalDetails.getFpProductId()+" deleted data count:"+deleteApplicationData);
